refactor viewer state into its own file
