+++ napari_animation/_tests/conftest.py
@@ -5,8 +5,7 @@ https://docs.pytest.org/en/stable/fixture.html
 import numpy as np
 import pytest
 
-from napari_animation import Animation
-from napari_animation.key_frame import ViewerState
+from napari_animation import Animation, ViewerState
 
 
 @pytest.fixture
